Squashed commit of the following:
commit 34ea660e7dcf0429414a7adf22edeab3a9de24d2
Author: AlgebraicJulia Bot <129184742+algebraicjuliabot@users.noreply.github.com>
Date:   Mon Apr 24 10:39:01 2023 -0400

    Set version to 0.2.0

commit f80c5136d79d53c4b11756c8667665723eef6d41
Author: Micah Halter <micah@mehalter.com>
Date:   Fri Apr 21 09:49:04 2023 -0400

    Remove architecture specific integer types

commit 3b3ab555498dce0eec3f28f0f255ca111cfaec7d
Author: Luke Morris <70283489+lukem12345@users.noreply.github.com>
Date:   Mon Apr 10 14:55:55 2023 -0400

    Remove NamedDecapode constructor and compiler

commit d2d5cfb13a52d65bc2f092b4b5beb8b03750ec51
Author: Luke Morris <70283489+lukem12345@users.noreply.github.com>
Date:   Mon Apr 10 14:44:44 2023 -0400

    Use @decapode macro

commit 28d6eb57eadb74f7ffb228bda1cb77318cb5cd08
Author: Micah Halter <micah@mehalter.com>
Date:   Mon Mar 27 10:36:20 2023 -0400

    Move to organization level github actions

commit b3c2476d585e1a97245c470ebba17776eb2897b5
Author: Luke Morris <70283489+lukem12345@users.noreply.github.com>
Date:   Mon Mar 20 14:38:08 2023 -0400

    Remove Decapodes_v0.1 files

+++ src/rewrite.jl
@@ -79,10 +79,10 @@ function preprocess_average_rewrite(deca_source::SummationDecapode)
     RHS = []
 
     SuperMatch = []
-    SuperVarMap = Vector{Int64}()
-    SuperOp2Map = Vector{Int64}()
-    SuperSigmaMap = Vector{Int64}()
-    SuperSummandMap = Vector{Int64}()
+    SuperVarMap = Vector{Int}()
+    SuperOp2Map = Vector{Int}()
+    SuperSigmaMap = Vector{Int}()
+    SuperSummandMap = Vector{Int}()
 
     serial = 0
     # Process all of the target rewrites for op2
@@ -227,8 +227,8 @@ function process_average_rewrite(deca_source::SummationDecapode)
     RHS = []
 
     SuperMatch = []
-    SuperVarMap = Vector{Int64}()
-    SuperOp1Map = Vector{Int64}()
+    SuperVarMap = Vector{Int}()
+    SuperOp1Map = Vector{Int}()
 
     varSerial = 0
     sumSerial = 0
